pyproject.toml
[tool.poetry]
name = "fs_flowstate_svc"
version = "1.0.0"
description = ""
authors = ["AIdev"]
readme = "README.md"
packages = [{include = "fs_flowstate_svc", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
alembic = "^1.14.0"
sqlalchemy = "^2.0.36"
pydantic = "^2.10.2"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
pydantic-settings = "^2.11.0"
bcrypt = "^5.0.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
email-validator = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^25.9.0"
isort = "^6.0.1"
mypy = "^1.18.2"
httpx = "^0.28.1"

[tool.poetry.scripts]
fs_flowstate_svc = "fs_flowstate_svc.main:main"

[tool.pytest.ini_options]
pythonpath = [ "src/" ]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"